1-ETL/pyscripts/agg_pull/requirements.txt
pandas>=1.1.2
bs4
numpy
polygon-api-client
pandas-market-calendars
gcsfs
google-cloud-storage
lxml
fsspec



